Still looking for a solution to fix the bug with circle

+++ include/circle.h
@@ -9,13 +9,13 @@
 #define CIRCLE_H
 
 typedef struct circle {
-	double x;
-	double y;
-	double radius;
+	float x;
+	float y;
+	float radius;
 } circle_t;
 
 void draw_circle(framebuffer_t *buffer, circle_t circle, sfColor color);
-void set_circle_pos(circle_t *circle, double x_new_pos, double y_new_pos);
+void set_circle_pos(circle_t *circle, float x_new_pos, float y_new_pos);
 
 void move_back(int *x, float speed);
 void move_forward(int *x, float speed);
